Move AppConfigurationException to scheduler module (#1196)
* Remove useless JobTypeStatistics

* Move AppConfigurationException to scheduler module

* Remove useless EmbedTestingServer
